Move `relay/conn.rs` to `relay/client/conn.rs` and move `http/client.rs` one level up
