+++ iroh-net/src/relay/client/conn.rs
@@ -16,9 +16,9 @@ use tokio_tungstenite_wasm::WebSocketStream;
 use tokio_util::codec::{FramedRead, FramedWrite};
 use tracing::{debug, info_span, trace, Instrument};
 
-use super::codec::PER_CLIENT_READ_QUEUE_DEPTH;
-use super::http::streams::{MaybeTlsStreamReader, MaybeTlsStreamWriter};
-use super::{
+use crate::relay::codec::PER_CLIENT_READ_QUEUE_DEPTH;
+use crate::relay::http::streams::{MaybeTlsStreamReader, MaybeTlsStreamWriter};
+use crate::relay::{
     codec::{
         write_frame, DerpCodec, Frame, MAX_PACKET_SIZE, PER_CLIENT_SEND_QUEUE_DEPTH,
         PROTOCOL_VERSION,
